+++ src/Elements/Contribuicoes/C100.php
@@ -2,7 +2,7 @@
 
 namespace NFePHP\EFD\Elements\Contribuicoes;
 
-use NFePHP\Common\Keys;
+use NFePHP\EFD\Common\ChaveAcesso;
 use NFePHP\EFD\Common\Element;
 use stdClass;
 
@@ -26,7 +26,7 @@ class C100 extends Element
 
     public function postValidation()
     {
-        if (!empty($this->std->chv_nfe) and !Keys::isValid($this->std->chv_nfe)) {
+        if (!empty($this->std->chv_nfe) and !ChaveAcesso::valida((string) $this->std->chv_nfe)) {
             $this->errors[] = "[" . self::REG . "] " .
                 " Dígito verificador incorreto no campo campo chave da " .
                 "nota fiscal eletronica (CHV_NFE)";